project restructure.
- matches desired submission structure
