+++ go_chase_it/ball_chaser/CMakeLists.txt
@@ -38,6 +38,3 @@ add_dependencies(drive_bot ball_chaser_generate_messages_cpp)
 add_executable(process_image src/process_image.cpp)
 target_link_libraries(process_image ${OpenCV_LIBRARIES} ${catkin_LIBRARIES})
 add_dependencies(process_image ball_chaser_generate_messages_cpp)
-
-add_executable(image_converter src/image_converter.cpp)
-target_link_libraries(image_converter ${OpenCV_LIBRARIES} ${catkin_LIBRARIES})
