+++ go_chase_it/ball_chaser/package.xml
@@ -6,7 +6,7 @@
 
   <maintainer email="andrew@todo.todo">andrew</maintainer>
 
-  <license>TODO</license>
+  <license>MIT</license>
 
   <buildtool_depend>catkin</buildtool_depend>
